+++ app/controllers/rails_omnibar/base_controller.rb
@@ -1,2 +1,5 @@
 class RailsOmnibar::BaseController < ActionController::API
+  def omnibar
+    @omnibar ||= params.fetch('omnibar_class').constantize
+  end
 end