+++ app/controllers/rails_omnibar/html_controller.rb
@@ -1,0 +1,5 @@
+class RailsOmnibar::HtmlController < RailsOmnibar::BaseController
+  def show
+    render html: omnibar.render
+  end
+end