+++ config/routes.rb
@@ -1,4 +1,5 @@
 RailsOmnibar::Engine.routes.draw do
   resource :js,    only: :show, defaults: { format: :js }
+  resource :html,  only: :show
   resource :query, only: :show
 end