+++ lib/rails_omnibar/command/base.rb
@@ -5,16 +5,16 @@ class RailsOmnibar
 
       def initialize(pattern:, resolver:, description: nil, example: nil)
         @pattern = cast_to_pattern(pattern)
-        @resolver = cast_to_proc(resolver, 2)
+        @resolver = RailsOmnibar.cast_to_proc(resolver)
         @description = description
         @example = example
       end
 
-      def call(input, omnibar = nil)
+      def call(input, controller:, omnibar:)
         match = pattern.match(input)
         # look at match for highest capturing group or whole pattern
         value = match.to_a.last || raise(ArgumentError, 'input !~ pattern')
-        results = resolver.call(value, omnibar)
+        results = resolver.call(value, controller: controller, omnibar: omnibar)
         results = results.try(:to_ary) || [results]
         results.map { |e| RailsOmnibar.cast_to_item(e) }
       end
@@ -31,13 +31,6 @@ class RailsOmnibar
           end
 
         regexp
-      end
-
-      def cast_to_proc(arg, arity)
-        arg = arg.is_a?(Proc) ? arg : arg.method(:call).to_proc
-        arg.is_a?(Proc) && arg.arity == arity ||
-          raise(ArgumentError, "arg must be a Proc with arity #{arity}")
-        arg
       end
     end
   end