+++ lib/rails_omnibar/item/help.rb
@@ -1,7 +1,6 @@
 class RailsOmnibar
   def add_help(**kwargs)
     add_item Item::Help.new(for_commands: commands, **kwargs)
-    self.class
   end
 
   module Item