+++ spec/lib/rails_omnibar/config_spec.rb
@@ -1,7 +1,9 @@
 require 'rails_helper'
 
+TestBar = Class.new(RailsOmnibar)
+
 describe RailsOmnibar do
-  subject { Class.new(RailsOmnibar).configure }
+  subject { TestBar }
 
   it 'has a configurable max_results' do
     expect(subject.max_results).to eq 10
@@ -49,5 +51,12 @@ describe RailsOmnibar do
     expect(subject.placeholder).to eq 'Hint: Type `REEE` for help'
     subject.placeholder = false
     expect(subject.placeholder).to eq nil
+  end
+
+  it 'raises when trying to configure or render from an anonymous class' do
+    klass = Class.new(RailsOmnibar)
+    expect { klass.configure {} }.to raise_error(/constant/)
+    expect { klass.add_item(title: 'a', url: 'b') }.to raise_error(/constant/)
+    expect { klass.render }.to raise_error(/constant/)
   end
 end