+++ spec/templates/app_template.rb
@@ -1,12 +1,18 @@
 gem 'rails_omnibar', path: __dir__ + '/../'
+gem 'devise'
 
 generate 'model', 'User first_name:string last_name:string admin:boolean --no-test-framework'
+generate 'devise:install'
+generate 'devise User'
 
-file 'app/lib/my_omnibar.rb', File.read(__dir__ + '/my_omnibar_template.rb')
+prepend_to_file 'config/application.rb', "require 'devise'\n"
+
+file 'app/lib/my_omnibar.rb',    File.read(__dir__ + '/my_omnibar_template.rb')
+file 'app/lib/other_omnibar.rb', File.read(__dir__ + '/other_omnibar_template.rb')
 
 inject_into_class 'app/controllers/application_controller.rb', 'ApplicationController', <<-RUBY
   def index
-    render html: MyOmnibar.render
+    render html: (MyOmnibar.render + OtherOmnibar.render)
   end
 RUBY