+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2023-01-31
+
+### Added
+
+- `RailsOmnibar::auth=` for fine-grained authorization
+- `RailsOmnibar::html_url` for rendering in SPAs
+
+### Fixed
+
+- double execution of commands / queries
+
 ## [1.3.2] - 2023-01-27
 
 ### Fixed